% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{impPlotsByChr}
\alias{impPlotsByChr}
\title{Graph imputed metrics: by chromosome boxplots}
\usage{
impPlotsByChr(out_dir, start_chr, end_chr, parX = FALSE, mets_all)
}
\arguments{
\item{out_dir}{Output directory}

\item{start_chr}{Start chromosome (1:23)}

\item{end_chr}{End chromosome (1:23)}

\item{parX}{logical whether to include PAR (not implemented)}

\item{mets_all}{Dataframe of imputation metrics created by \code{\link{readMetrics}}}
}
\value{
Boxplots by chromosome of MAF and Rsq
}
\description{
Graph imputed metrics: by chromosome boxplots
}
