% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{impPlotsInfo}
\alias{impPlotsInfo}
\title{Graph imputed metrics: Rsq by MAF}
\usage{
impPlotsInfo(out_dir, start_chr, end_chr, parX = FALSE, fmt = "pdf",
  mets_all)
}
\arguments{
\item{out_dir}{Output directory}

\item{start_chr}{Start chromosome (1:23)}

\item{end_chr}{End chromosome (1:23)}

\item{parX}{logical whether to include PAR (not implemented)}

\item{fmt}{Plot output format (only pdf supported)}

\item{mets_all}{Dataframe of imputation metrics created by \code{\link{readMetrics}}}
}
\value{
Plot of Rsq by MAF in imputed variants
}
\description{
Graph imputed metrics: Rsq by MAF
}
