% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{impSummary}
\alias{impSummary}
\title{Create counts of imputation basis and target variants}
\usage{
impSummary(start_chr, end_chr, parX = FALSE, out_dir, mets_all,
  keep_list = NA)
}
\arguments{
\item{start_chr}{Start chromosome (1:23)}

\item{end_chr}{End chromosome (1:23)}

\item{parX}{logical whether to include PAR (not implemented)}

\item{out_dir}{Output directory}

\item{mets_all}{Dataframe of imputation metrics created by \code{\link{readMetrics}}}

\item{keep_list}{File path of an initial variant keep list, with columns variant name and chromosome (integer)}
}
\value{
Writes a csv file of variant counts
}
\description{
Create counts of imputation basis and target variants
}
