% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{makeAllPlots}
\alias{makeAllPlots}
\title{Wrapper function to run all plotting and summary functions}
\usage{
makeAllPlots(imp_dir, out_dir, start_chr = 1, end_chr = 23,
  sets = c("imputed", "masked"), plots = c("boxplot", "info_by_maf",
  "masked_check_strand"), fmt = "pdf", summary = TRUE, rsq_filt = 0.8,
  verbose = TRUE, maf_thresh = 0.05, parX = FALSE, keep_list = NA)
}
\arguments{
\item{imp_dir}{Imputation directory}

\item{out_dir}{Output directory}

\item{start_chr}{Start chromosome (1:23)}

\item{end_chr}{End chromosome (1:23)}

\item{sets}{Sets of imputation summaries}

\item{plots}{Sets of plots}

\item{fmt}{Plot output format (only pdf supported)}

\item{summary}{Logical indicator of whether to run \code{\link{impSummary}}}

\item{rsq_filt}{Rsq threshold used in \code{\link{maskedPlots}}}

\item{verbose}{Logical indicator for verbose reporting}

\item{maf_thresh}{Minor allele frequency threshold for grouping masked variants in \code{\link{maskedSummary}}}

\item{parX}{logical whether to include PAR (not implemented)}

\item{keep_list}{File path of an initial variant keep list, with columns variant name and chromosome (integer)}
}
\description{
Wrapper function to run all plotting and summary functions
}
