% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{maskedPlots}
\alias{maskedPlots}
\title{Graph masked variant  metrics}
\usage{
maskedPlots(out_dir, start_chr, end_chr, parX = FALSE, rsq_filt, mets_all)
}
\arguments{
\item{out_dir}{Output directory}

\item{start_chr}{Start chromosome (1:23)}

\item{end_chr}{End chromosome (1:23)}

\item{parX}{logical whether to include PAR (not implemented)}

\item{rsq_filt}{Rsq threshold for binning masked variants}

\item{mets_all}{Dataframe of imputation metrics created by \code{\link{readMetrics}}}
}
\value{
4x4 plot of masked variant metrics
}
\description{
Graph masked variant  metrics
}
