% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{maskedStrandCheck}
\alias{maskedStrandCheck}
\title{Plot masked variant strand check}
\usage{
maskedStrandCheck(out_dir, start_chr, end_chr, mets_all)
}
\arguments{
\item{out_dir}{Output directory}

\item{start_chr}{Start chromosome (1:23)}

\item{end_chr}{End chromosome (1:23)}

\item{mets_all}{Dataframe of imputation metrics created by \code{\link{readMetrics}}}
}
\value{
Scatterplot of masked variant metrics that can help identify systematic strand alignment issues
Caveat the resulting plot can be messy. Read the log file to understand the count of variants in the problematic quadrant: those imputed with high confidence but low concordance.
}
\description{
Plot masked variant strand check
}
