% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{readMetrics}
\alias{readMetrics}
\title{Read in imputation metrics files}
\usage{
readMetrics(imp_dir, project, start_chr, end_chr, parX = FALSE)
}
\arguments{
\item{imp_dir}{Imputation directory containing metrics files}

\item{project}{Project name}

\item{start_chr}{Start chromosome (1:23)}

\item{end_chr}{End chromosome (1:23)}

\item{parX}{logical whether to include PAR (not implemented)}
}
\value{
Dataframe of imputation metrics, from \code{start_chr} to \code{end_chr}
}
\description{
Reads in metrics files generated by University of Michigan Imputation Server. 
Note: optimized for use with reference HRC 1.1 (SNP only)
}
\examples{
\dontrun{
mets <- readMetrics("mymetsdir", "proj", start_chr=1, end_chr=23)
}
}
